Deal with first line of blocks in VB
These are the first line of for loops, while loops, select statements, etc.

Basically make sure we've visited the containing first line of such a construct before porting any child trivia. And when visiting the direct parent of that block, repoint trivia at the end of the block

This is basically an extension of the idea of avoiding porting trivia until the token isn't the last token of its parent.

Issues remaining:

- IsFirstLineOfBlockConstruct isn't quite accurate, it probably shouldn't pick up one line if statements, or at least shouldn't port trivia from the end of them into the middle
- Over-removal of annotationData is somehow happening, which means some updatedtokens are being lost, potentially due to misunderstanding of the replacetokens function parameters
- Backporting is a bit overzealous for porting all trivia from if else block right back to the start, and moving ahead of semicolon in a few cases (with block and empty while block)