Allow our extension to load any version of the assemblies lying around
This only kicks in when the runtime is already having difficulty finding the required assembly anyway.
This way we be sure of compatibility with the latest version if we compile against the latest assembly shipped with VS, then just hope for compatibility with older versions.

The alternative is to compile against the oldest versions of the assemblies for a version of VS we care about, and rely on Visual Studio's own binding redirects (or our advice to it in assemblyinfo) to sort things out. The APIs we're using are quite stable, but I'd rather old versions break than new versions since "please upgrade VS" is a better response to a bug report than "please downgrade VS". If we get lots of MissingMethodException bug reports, we'll have to reconsider this.