Use annotations to keep track of the node to be converted.
The benefit over converting just the selected characters is that we're converting what's hopefully a well formed document, so the rest of the document can be used for context, and returning something that represents a node, rather than trying to convert the text "tring", when someone slightly mis-selects.

If two layers of the node hierarachy fail to go through the trivia converter, then the whole document will be converted, which seems like an OK failure case.