Use AsyncServiceProvider
Still need to get the right interop type, and figure out how to show a message box using the IAsyncServiceProvider (can I get a non-async service provider from it to pass to the message box function? Or do I need to find a different method call entirely to do this? Check if other extensions like RefactoringEssentials have already tackled this)