Handle errors in type members
For c#, I've created a dummy class to attach the error comment to, since empty statements are invalid in many contexts in C#, and there may be no previous node to attach to. Attaching to a parent would result in the code being out of order from where it was originally, so this seems best.