Add some more failing test cases
I suspect the issue isn't that the control flow analysis doesn't understand +=
It's that AlwaysAssigned is telling you the result between two code points.
Really what we need to know, is what's the first statement that reads the variable, and is it alwaysassigned by that point.
Whether you include the statement itself in that range depends whether the first access that statement makes to the variable is a read or a write