Use declared accessibility rather than trying to derive it from syntax
I've left in some special cases where the defaults are the same between the language so that omitting a keyword has the same effect in the conversion in some cases where that has no semantic effect.
This will likely be the source of later bugs - it's probably worth changing to just always be explicit at some point (except places where it's illegal to do so, e.g. a property getter/setter with the same visibility as its parent property)