Ensure references are added to the default set - not replacing them
And make sure Enumerable's assembly is referenced

Relates to https://github.com/icsharpcode/CodeConverter/issues/342

The tests weren't seeing the issue since they end up running in a different runtime than the web app - types have moved between assemblies a lot
It'd be possible to sync those two up, but the VS extension can be loaded into tons of different runtimes depending on VS version.
It might be possible to set the framework version for a compilation and then pick a more static set of assemblies, but that will cause issues when people don't have that particular targeting pack.
For now, this seems to fix the visible issue