Revert usage of syntax generator
It incorrectly generated "static const" which the converter can already do better.
I'll keep it available for future use though. It might make sense to build it in as the core of each declaration with other bits added on.
The main issue with heavier use of ISymbol and IOperation is that it will likely be less effective on slightly broken compilations.
At some point it may be necessary to split off some of the current syntax-focussed code to handle just those cases.