Give VisualStudio hints about context menu visibility before load
This seems to work well for: The edit toolbar menu, the editor context menu, inidividual cs and vb files.
This partially works for: csproj and vbproj. It only works for a small (seemingly arbitrary) fraction of them
This never works for: The sln file

Some research into the exact meaning of