Only convert important trivia
The issue is that the last node in a statement differs in vb syntax.
For comments it's crucial to port trivia for statements and declarations, any other comments can be moved around.
For newlines it's crucial not to port the newline at the end of a statement to the equivalent node, it must be (if at all) ported to the equivalent StatementSyntax

It might be easiest to refocus and just port comment trivia everywhere