Tell users about this new switch when it inevitably causes pain
Would be better off doing some simple detection on the project files probably.
The framework version of the exe is actually more flexible since it works for framework and core, but it won't work on other platforms. It may make sense to just try to detect when the framework one is runnable.