Regex replace
\(([A-Za-z]*)\)\s?await (.*).AcceptAsync\(
await $2.AcceptAsync<$1>(

The intent here is to push the information about expected type to somewhere that can do something useful with the information in a more central place.
You could also imagine having "default" elements for a bunch of common types (e.g. identifiernamesyntax) that allow us to attach an error in a more specific context
To go even further, the return type of the visitor could contain multiple options, and the AcceptAsync method pattern picking the one relevant to the context